% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_ts.R
\name{animate_ts}
\alias{animate_ts}
\title{Animate Time-Series Data}
\usage{
animate_ts(
  df,
  x_col = "date",
  y_col = "value",
  color = "blue",
  title = "Time Series Animation"
)
}
\arguments{
\item{df}{(Dataframe) Time-series data from \code{fetch_ts()}.}

\item{x_col}{(Character) Name of the column containing timestamps (default: "date").}

\item{y_col}{(Character) Name of the column with values to plot (default: "value").}

\item{color}{(Character) Line color (default: "blue").}

\item{title}{(Character) Plot title (default: "Time Series Animation").}
}
\value{
A \code{gganimate} object.
}
\description{
Creates a customizable animation using \code{ggplot2} and \code{gganimate}.
}
\examples{
\dontrun{
ndvi_df<- fetch_ts(
  collection = "LANDSAT/LC08/C01/T1_8DAY_NDVI",
  band = "NDVI",
  region = region,
  dates = c("2010-01-01", "2020-12-31"),
  scale = 30
)
anim <- animate_ts(ndvi_df, y_col = "NDVI", color = "red", title = "NDVI Trends( 2010-2020")
render_animation(anim, "ndvi_trends.gif")
}
}
