% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ts.R
\name{fetch_ts}
\alias{fetch_ts}
\title{Fetch Time-Series Data from Google Earth Engine}
\usage{
fetch_ts(collection, band, region, dates, reducer = "mean", scale = 30)
}
\arguments{
\item{collection}{GEE dataset ID. Options:
\itemize{
\item Landsat NDVI: "LANDSAT/LC08/C01/T1_8DAY_NDVI"
\item MODIS NDVI: "MODIS/006/MOD13A2"
\item Sentinel-2: "COPERNICUS/S2_SR"
}}

\item{band}{Band name (e.g., "NDVI", "B4", "NDWI").}

\item{region}{Spatial region to analyze (sf object for EE Geometry).}

\item{dates}{Start and end dates in "YYYY-MM-DD" format..}

\item{reducer}{Statistic to compute over the region: "mean", "median", "min", "max" (default: "mean").}

\item{scale}{Pixel resolution in meters (e.g., 30 for Landsat, 500 for MODIS).}
}
\value{
A dataframe with columns 'date' (timestamp) and 'value' (extracted band values).
}
\description{
Extracts a time series of a specified band over a region from a GEE collection.
}
\examples{
\dontrun{
# Get NDVI from Landsat 8
library(sf)
gee_init()
region <- st_as_sfc("POLYGON ((-110.8 30.2, -110.6 30.2, -110.6 30.4, -110.8 30.4, -110.8 30.2))")
ndvi_df<- fetch_ts(
  collection = "LANDSAT/LC08/C01/T1_8DAY_NDVI",
  band = "NDVI",
  region = region,
  dates = c("2010-01-01", "2020-12-31"),
  scale = 30
)
head(ndvi_data)
}
}
